Updated wrapper to the new Tox API.
This commit updates the wrapper to the new API. Clients using this
wrapper will need to upgrade some code. To be more consistent with the C
interface, some functions where renamed.

This update also fixes some bugs causing the wrapper to crash.

Sending files and groupchats are currently not supported.